refactor(cli): split main.rs into focused modules
Split the God File main.rs (1,347 lines) into smaller, focused modules:

- cli.rs: CLI definitions (Cli, Commands, RulesCommands, PluginCommands, CacheCommands enums)
- commands/mod.rs: Command module entry point
- commands/lint.rs: run_lint, find_config logic
- commands/init.rs: run_init logic
- commands/rules.rs: run_create_rule, run_add_rule logic
- commands/plugin.rs: run_plugin_install, run_plugin_cache_clean logic
- commands/lsp.rs: run_lsp logic
- output/mod.rs: Output formatting module
- output/text.rs: Text formatter
- output/json.rs: JSON formatter
- output/sarif.rs: SARIF formatter
- config/editor.rs: update_config_with_plugin, ConfigEdit logic
- fix.rs: FixSummary, apply_fixes, output_fix_summary

main.rs is now a thin entry point (93 lines) that delegates to modules.
