+++ crates/tsuzulint_cli/src/commands/mod.rs
@@ -1,0 +1,7 @@
+//! Command implementations
+
+pub mod init;
+pub mod lint;
+pub mod lsp;
+pub mod plugin;
+pub mod rules;
