+++ crates/tsuzulint_cli/src/config/mod.rs
@@ -1,0 +1,3 @@
+//! Configuration module
+
+pub mod editor;
